requirements.txt
kucoin-futures-python
python-dotenv
thread6
asyncio